don't use the lock facility.
In some circumstances there are many processes of ulockmgr created and never
killed again. this causes the exhaustion of filedescriptors and some error msg
like "Too many open files."
Don't use it Since the ciopfs_lock function is just the copy of the example and
the hint in the fuse-documentation that states:

    Note: if this method is not implemented, the kernel will still allow file
    locking to work locally. Hence it is only interesting for network
    filesystems and similar.

I don't think that someone knows what it should do either.
If somebody knows, please send a patch.